Refactor IConfig file into its own package
