+++ animatedarchitecture-core/src/main/java/nl/pim16aap2/animatedarchitecture/core/animation/AnimationRequestData.java
@@ -6,10 +6,10 @@ import dagger.assisted.AssistedInject;
 import jakarta.inject.Named;
 import lombok.Getter;
 import nl.pim16aap2.animatedarchitecture.core.api.HighlightedBlockSpawner;
-import nl.pim16aap2.animatedarchitecture.core.api.IConfig;
 import nl.pim16aap2.animatedarchitecture.core.api.IExecutor;
 import nl.pim16aap2.animatedarchitecture.core.api.IPlayer;
 import nl.pim16aap2.animatedarchitecture.core.audio.IAudioPlayer;
+import nl.pim16aap2.animatedarchitecture.core.config.IConfig;
 import nl.pim16aap2.animatedarchitecture.core.events.StructureActionCause;
 import nl.pim16aap2.animatedarchitecture.core.events.StructureActionType;
 import nl.pim16aap2.animatedarchitecture.core.managers.AnimationHookManager;
