+++ animatedarchitecture-core/src/main/java/nl/pim16aap2/animatedarchitecture/core/config/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This package contains all core classes of the configuration system.
+ */
+@NonNullByDefault
+package nl.pim16aap2.animatedarchitecture.core.config;
+
+import org.eclipse.jdt.annotation.NonNullByDefault;
