+++ animatedarchitecture-core/src/main/java/nl/pim16aap2/animatedarchitecture/core/managers/LimitsManager.java
@@ -2,9 +2,9 @@ package nl.pim16aap2.animatedarchitecture.core.managers;
 
 import jakarta.inject.Inject;
 import jakarta.inject.Singleton;
-import nl.pim16aap2.animatedarchitecture.core.api.IConfig;
 import nl.pim16aap2.animatedarchitecture.core.api.IPermissionsManager;
 import nl.pim16aap2.animatedarchitecture.core.api.IPlayer;
+import nl.pim16aap2.animatedarchitecture.core.config.IConfig;
 import nl.pim16aap2.animatedarchitecture.core.util.Limit;
 
 import java.util.OptionalInt;
