+++ animatedarchitecture-core/src/main/java/nl/pim16aap2/animatedarchitecture/core/util/delayedinput/DelayedStructureSpecificationInputRequest.java
@@ -6,10 +6,10 @@ import lombok.CustomLog;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
 import lombok.experimental.ExtensionMethod;
-import nl.pim16aap2.animatedarchitecture.core.api.IConfig;
 import nl.pim16aap2.animatedarchitecture.core.api.IExecutor;
 import nl.pim16aap2.animatedarchitecture.core.api.ILocation;
 import nl.pim16aap2.animatedarchitecture.core.api.IPlayer;
+import nl.pim16aap2.animatedarchitecture.core.config.IConfig;
 import nl.pim16aap2.animatedarchitecture.core.localization.PersonalizedLocalizer;
 import nl.pim16aap2.animatedarchitecture.core.managers.StructureSpecificationManager;
 import nl.pim16aap2.animatedarchitecture.core.structures.Structure;
