+++ animatedarchitecture-testing/animatedarchitecture-integration-test/src/test/java/nl/pim16aap2/animatedarchitecture/creator/CreatorTestsUtil.java
@@ -3,7 +3,6 @@ package nl.pim16aap2.animatedarchitecture.creator;
 import nl.pim16aap2.animatedarchitecture.core.UnitTestUtil;
 import nl.pim16aap2.animatedarchitecture.core.animation.StructureActivityManager;
 import nl.pim16aap2.animatedarchitecture.core.api.IAnimatedArchitectureToolUtil;
-import nl.pim16aap2.animatedarchitecture.core.api.IConfig;
 import nl.pim16aap2.animatedarchitecture.core.api.IEconomyManager;
 import nl.pim16aap2.animatedarchitecture.core.api.IExecutor;
 import nl.pim16aap2.animatedarchitecture.core.api.IPermissionsManager;
@@ -20,6 +19,7 @@ import nl.pim16aap2.animatedarchitecture.core.commands.DelayedCommandInputReques
 import nl.pim16aap2.animatedarchitecture.core.commands.SetBlocksToMoveDelayed;
 import nl.pim16aap2.animatedarchitecture.core.commands.SetOpenDirectionDelayed;
 import nl.pim16aap2.animatedarchitecture.core.commands.SetOpenStatusDelayed;
+import nl.pim16aap2.animatedarchitecture.core.config.IConfig;
 import nl.pim16aap2.animatedarchitecture.core.localization.ILocalizer;
 import nl.pim16aap2.animatedarchitecture.core.managers.DatabaseManager;
 import nl.pim16aap2.animatedarchitecture.core.managers.DelayedCommandInputManager;
